add a custom source code lister and make the images "file format quines"
the flat binaries are now pretty much valid html, too. that means that
if you open your image in a browser it will (hopefully) show you the
code that was used to build that binary. that also makes the images a
lot bigger and introduces many new hacks, but it's fun!

i also made a custom system for this, because it is such a simple
problem and getting doxygen to generate something useful seemed
intimidating.

i've been working on this for pretty much the past 8 hours.

first tried making it pdf+flatbin:
https://twitter.com/pixelcmtd/status/1526073851033042944; then
considered postscript:
https://twitter.com/pixelcmtd/status/1526075757352603648; and re-watched
an old talk to learn how to do it: https://youtu.be/hdCs6bPM4is; there i
got the idea for html and got it working somewhat quickly:
https://twitter.com/pixelcmtd/status/1526124801701687296